Refine the startup message for Web-based service
This change adds a notice to the startup message to warn users that
future checker releases will bind by default to 127.0.0.1 instead of
binding to all interfaces (by using 0.0.0.0). The change adds some code
that attempts to get the IP4 addresses for any open interfaces in the
environment, and then emits a message with examples showing how to bind
the checker to those addresses (instead of 0.0.0.0 or 127.0.0.1).